zeon/var/www/zeon/requirements.txt
alembic==1.4.3
apispec==3.3.2
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.1.0
click==7.1.2
cyrtranslit==1.0
flasgger==0.9.5
Flask==1.1.4
Flask-Login==0.5.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==2.10
importlib-metadata==2.1.3
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.6
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.18.0
mistune==2.0.2
psycopg2-binary==2.9.3
pyrsistent==0.17.3
python-dateutil==2.8.2
python-decouple==3.6
python-editor==1.0.4
PyYAML==5.3.1
pyzabbix==1.0.0
requests==2.25.1
semantic-version==2.10.0
six==1.16.0
SQLAlchemy==1.3.24
urllib3==1.26.9
Werkzeug==1.0.1
zipp==1.2.0
